requirements.txt
streamlit
google-cloud-speech